requirements.txt
anyio==4.7.0
asgiref==3.7.2
blinker==1.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
clang==17.0.6
click==8.2.1
cloudipsp==1.0.4
contourpy==1.3.1
cycler==0.12.1
dj-database-url==3.0.1
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.2
fonttools==4.55.3
fsspec==2024.12.0
gradio_client==1.5.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
Pillow==10.0.0
psycopg==3.1.9
psycopg2-binary==2.9.6
pygame==2.6.1
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.2
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.1
typing_extensions==4.6.3
urllib3==2.0.4
websockets==14.1
Werkzeug==3.1.3
